requirements_asr_app.txt
streamlit>=1.28.0
torch>=2.0.0
transformers>=4.35.0
peft>=0.6.0
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.24.0
accelerate>=0.24.0
faiss-cpu>=1.7.4
pandas>=1.5.0
sentence-transformers>=2.2.0
google-generativeai>=0.3.0
edge-tts>=6.1.0
gtts>=2.5.0
jiwer>=3.0.0

